+++ sources/ft_absolute_num.c
@@ -1,20 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_absolute.c                                      :+:      :+:    :+:   */
+/*   ft_absolute_num.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: adantas- <adantas-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/27 23:02:05 by adantas-          #+#    #+#             */
-/*   Updated: 2023/01/31 10:56:14 by adantas-         ###   ########.fr       */
+/*   Updated: 2023/04/17 16:33:45 by adantas-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-ssize_t	ft_absolute(ssize_t n)
+/**
+ * @brief Get the absolute value of a number if it is a negative;
+ * 
+ * @param number The number to be converted to absolute;
+ * @return The number in absolute form;
+**/
+ssize_t	ft_absolute_num(ssize_t number)
 {
-	if (n < 0)
-		return (-n);
-	return (n);
+	if (number < 0)
+		return (-number);
+	return (number);
 }
